File and comment cleanup
